Ensure user._id is string before comparison
Make sure the user's ID is a string before comparing with a value that we know is a string. Before it was a MongoDB object representing the ID, but it was not 100% a string. Now only a string *should* be able to be compared to determine if the user can see this information.